Adjust the speeds of a gps-like table created with gtfs2gps

Some GTFS.zip data sets might have quality issues, for example by assuming that a trip speed is unreasonably high (e.g. an urban bus running over 100 Km/h), or in other cases the `departure_time` information might be missing for some route segments. This can lead a gps-like table to have `NA` or unrealistic `speed` and `departure_time` values. This function allows the user to adjust the speed of trips and updates `departure_time` values accordingly. The user can adjust the problematic speeds by either setting a custom constant value, or by considering the average of all valid trips speed (Default). The columns `departure_time` and `cumtime` are updated accordingly.

adjust_speed(gps_data, min_speed = 2, max_speed = 80, new_speed = NULL)

Arguments

gps_data
A GPS-like data.table created with gtfs2gps.

min_speed
Minimum speed (in km/h) to be considered as valid. Values below minimum speed will be adjusted. Defaults to 2 km/h.

max_speed
Maximum speed (in km/h) to be considered as valid. Values above maximum speed will be adjusted. Defaults to 80 km/h.

new_speed
Speed (in km/h) to replace missing values as well as values outside min_speed and max_speed range. By default, `new_speed = NULL` and the function considers the average speed of the entire gps data.

Value

A GPS-like data with adjusted `speed` values. The columns `departure_time` and `cumtime` are also updated accordingly.

Examples

poa <- read_gtfs(system.file("extdata/poa.zip", package="gtfs2gps"))

#> Unzipped the following files to /tmp/RtmpLgLrZt/gtfsio:
#>   * agency.txt
#>   * calendar.txt
#>   * routes.txt
#>   * shapes.txt
#>   * stop_times.txt
#>   * stops.txt
#>   * trips.txt

#> Reading agency.txt

#> Reading calendar.txt

#> Reading routes.txt

#> Reading shapes.txt

#> Reading stop_times.txt

#> Reading stops.txt

#> Reading trips.txt

poa_gps <- gtfs2gps(poa)

#> Converting shapes to sf objects

#> Processing the data

#> 
#> Attaching package: ‘data.table’

#> The following object is masked from ‘package:purrr’:
#> 
#>     transpose

#> Linking to GEOS 3.6.2, GDAL 2.2.3, PROJ 4.9.3

#> 
#> Attaching package: ‘magrittr’

#> The following object is masked from ‘package:purrr’:
#> 
#>     set_names

#> udunits database from /usr/share/xml/udunits/udunits2.xml

#> 73 'speed' values are NA for shapeid '176-1'.

#> 30 'speed' values are NA for shapeid 'A141-1'.

#> 108 'speed' values are NA for shapeid 'R10-2'.

#> 1764 'speed' values are NA for shapeid 'T2-1'.

#> Some 'speed' values are NA in the returned data.

poa_gps_new <- adjust_speed(poa_gps)